+++ imap_tools/__init__.py
@@ -1,7 +1,8 @@
 from .query import AND, OR, NOT, Header, UidRange, A, O, N, H, U
 from .mailbox import BaseMailBox, MailBox, MailBoxUnencrypted
-from .message import MailMessage, MailAttachment, MailMessageFlags
-from .folder import MailBoxFolderManager, MailBoxFolderStatusOptions
+from .message import MailMessage, MailAttachment
+from .folder import MailBoxFolderManager
+from .consts import MailMessageFlags, MailBoxFolderStatusOptions
 from .errors import *
 
-__version__ = '0.41.0'
+__version__ = '0.42.0'