+++ imap_tools/folder.py
@@ -1,28 +1,10 @@
 import re
 
 from . import imap_utf7
+from .consts import MailBoxFolderStatusOptions
 from .utils import check_command_status, pairs_to_dict, encode_folder
 from .errors import MailboxFolderStatusValueError, MailboxFolderSelectError, MailboxFolderCreateError, \
     MailboxFolderRenameError, MailboxFolderDeleteError, MailboxFolderStatusError
-
-
-class MailBoxFolderStatusOptions:
-    """Valid mailbox folder status options"""
-    MESSAGES = 'MESSAGES'
-    RECENT = 'RECENT'
-    UIDNEXT = 'UIDNEXT'
-    UIDVALIDITY = 'UIDVALIDITY'
-    UNSEEN = 'UNSEEN'
-    all = (
-        MESSAGES, RECENT, UIDNEXT, UIDVALIDITY, UNSEEN
-    )
-    description = (
-        (MESSAGES, "The number of messages in the mailbox"),
-        (RECENT, "The number of messages with the Recent flag set"),
-        (UIDNEXT, "The next unique identifier value of the mailbox"),
-        (UIDVALIDITY, "The unique identifier validity value of the mailbox"),
-        (UNSEEN, "The number of messages which do not have the Seen flag set"),
-    )
 
 
 class MailBoxFolderManager: